Move redux connect from Candidate's component to Jobs container
